refactor(masterkey): move internal/masterkey to pkg/masterkey
Pure relocation so out-of-tree consumers (viperblock) can import the
loader. Internal/ visibility is package-tree-relative and Go does not
support workspace-level allowlists; vendoring would duplicate a
trust-critical loader. Public API unchanged: Load, LoadShared, NewAEAD,
Fingerprint, Key, MasterKeySize all stay exported with identical
signatures.

Five import sites updated (s3/auth.go, s3/server.go,
quic/quicserver/server.go, internal/storetest/keys.go, the test file
itself). One doc comment in storetest/keys.go updated to reference the
new path.

Prerequisite for viperblock encryption-at-rest (mulga-bm-29..33).
